+++ frontend/src/pages/oj/views/user/learnerNotebookState.js
@@ -1,8 +1,0 @@
-export function toggleExpandedGroup (expandedGroups, key) {
-  const current = Boolean(expandedGroups && expandedGroups[key])
-  return Object.assign({}, expandedGroups, { [key]: !current })
-}
-
-export function forceExpandGroup (expandedGroups, key) {
-  return Object.assign({}, expandedGroups, { [key]: true })
-}